Make deep-merge the default behaviour and fix mutability bug in it
